Option to prevent menu close on itemclick
I would like to have the option to prevent menu closing when clicking on items, for example in case of using toggles it's annoying the menu closes on every click. Current PR only focuses on the `drop-down-icon` component, however, if useful, may be applied to other menu components as well.