Add 'swipe away' functionality to the left / right nav
Resolves #411.

If a user opens the nav on a touch device, they can use the swipe
gesture they're used to to close the nav.

Technical details: I initially tried setting nav's translateX and
overlay's opacity using setState, but turns out that on mobile devices
re-rendering the component on every touchmove event is costly and
leads to laggy animations.
So I resorted to directly manipulating elements' styles, which is
a lot faster.